[Update] Getter for the data state of a simulation
Major:
- This is needed to display information stored in the Data State in the Model Explorer

Minor:
- Fixed a precision warning

+++ ECell_Engine/src/SBMLModuleImporter.cpp
@@ -164,7 +164,7 @@ Operation ECellEngine::IO::SBMLModuleImporter::ASTNodeToOperation(
             if (cNode->getType() == ASTNodeType_t::AST_INTEGER ||
                 cNode->getType() == ASTNodeType_t::AST_REAL)
             {
-                pOP->AddConstant(cNode->getValue());
+                pOP->AddConstant((float)cNode->getValue());
                 nodesStack.pop_back();
             }
